Fixed Alerts

+++ microservices/backend/delete-ms/src/index.js
@@ -22,6 +22,11 @@ app.use(express.json());
 
 // Functions
 async function deleteUser(req, res) {
+  if(mongoose.connection.readyState == 0){
+    res.status(504).json({ error: 'Database is not connected' });
+    console.log('Database is not connected');
+    return;
+  }
   const { idNumber } = req.body;
   try {
     const user = await User.findOneAndDelete({ idNumber: idNumber });
@@ -43,6 +48,9 @@ async function deleteUser(req, res) {
 
 // Routes
 app.delete('/', deleteUser);
+app.get('/status', (req, res) => {
+  res.status(200).send('Server is running!');
+});
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found.' });
 });
